requirements.txt
absl-py==1.4.0
appdirs==1.4.4
async-generator==1.10
attrs==22.2.0
axisregistry==0.3.11
babelfont==3.0.0
beautifulsoup4==4.11.2
beziers==0.5.0
blackrenderer==0.6.0
booleanOperations==0.9.0
Brotli==1.0.9
bump2version==1.0.1
bumpfontversion==0.4.1
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
cffsubr==0.2.9.post1
charset-normalizer==3.0.1
click==8.1.3
cmarkgfm==2022.10.27
collidoscope==0.4.1
colorlog==6.7.0
commandlines==0.4.1
compreffor==0.5.3
cu2qu==1.6.7.post2
defcon==0.10.2
dehinter==4.0.0
Deprecated==1.2.13
diffenator2==0.0.6
drawbot-skia==0.5.0
exceptiongroup==1.1.0
font-v==2.1.0
fontbakery==0.8.11
fontFeatures==1.7.4
fontmake==3.5.1
fontMath==0.9.3
fonttools==4.38.0
freetype-py==2.3.0
fs==2.4.16
gflanguages==0.4.4
gftools==0.9.26
gitdb==4.0.10
GitPython==3.1.31
glyphsets==0.5.4
glyphsLib==6.1.0
glyphtools==0.8.0
h11==0.14.0
hyperglot==0.4.4
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
munkres==1.1.4
nanoemoji==0.15.1
ninja==1.11.1
numpy==1.24.2
openstep-plist==0.3.0.post1
opentype-sanitizer==9.0.0
opentypespec==1.9.1
orjson==3.8.7
outcome==1.2.0
packaging==23.0
picosvg==0.20.6
Pillow==9.4.0
pip-api==0.0.30
pngquant-cli==2.17.0.post5
protobuf==3.20.3
pyahocorasick==2.0.0
pybind11==2.10.3
pyclipper==1.3.0.post4
pycparser==2.21
pygit2==1.11.1
PyGithub==1.58.0
Pygments==2.14.0
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
resvg-cli==0.22.0.post3
rich==13.3.2
rstr==3.2.0
selenium==4.8.2
sh==2.0.2
six==1.16.0
skia-pathops==0.7.4
skia-python==87.5
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sre-yield==1.2
statmake==0.6.0
strictyaml==1.6.2
stringbrewer==0.0.1
tabulate==0.9.0
toml==0.10.2
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.9.2
ttfautohint-py==0.5.1
typing_extensions==4.5.0
ufo2ft==2.30.0
ufoLib2==0.14.0
ufolint==1.2.0
uharfbuzz==0.34.0
unicodedata2==15.0.0
Unidecode==1.3.6
urllib3==1.26.14
vharfbuzz==0.2.0
vttLib==0.11.0
wrapt==1.15.0
wsproto==1.2.0
youseedee==0.3.0
zipp==3.15.0
zopfli==0.2.2